updated inputs and gridsearch
we updated contact matrix input, type inputs. disease parameters are TBD but i extended the age bin to be consistent with the rest of inputs.
codes to generate contact matrices are updated accordingly
we have a new process to run gridsearch in 4b_grid_search
codes for SEIR are not updated and will not work